# docs/api/generated-swagger.yaml
openapi: 3.0.3
info:
  title: NextDocs API
  description: |-
    Auto-generated API documentation for NextDocs

    Generated on: 2025-12-14T09:37:21.234Z
  version: 1.0.0-20251214
  contact:
    name: NextDocs Team
servers:
  - url: /api
    description: API Server
security:
  - sessionAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: NextAuth.js session cookie
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
      description: API Key authentication (64-char hex string)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error description
      required:
        - error
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /admin/api-keys:
    get:
      tags:
        - Admin
      summary: GET /admin/api-keys
      description: Auto-generated documentation for GET /admin/api-keys
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Admin
      summary: POST /admin/api-keys
      description: Auto-generated documentation for POST /admin/api-keys
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
  /admin/api-keys/{id}:
    patch:
      tags:
        - Admin
      summary: PATCH /admin/api-keys/{id}
      description: Auto-generated documentation for PATCH /admin/api-keys/{id}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: id parameter
    delete:
      tags:
        - Admin
      summary: DELETE /admin/api-keys/{id}
      description: Auto-generated documentation for DELETE /admin/api-keys/{id}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: id parameter
...
  /admin/swagger:
    get:
      tags:
        - Admin
      summary: GET /admin/swagger
      description: Auto-generated documentation for GET /admin/swagger
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    options:
      tags:
        - Admin
      summary: OPTIONS /admin/swagger
      description: Auto-generated documentation for OPTIONS /admin/swagger
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /features/{id}/comments:
    get:
      tags:
        - Features
      summary: GET /features/{id}/comments
      description: Auto-generated documentation for GET /features/{id}/comments
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: id parameter
    post:
      tags:
        - Features
      summary: POST /features/{id}/comments
      description: Auto-generated documentation for POST /features/{id}/comments
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: id parameter
...